requirements.txt
streamlit==1.33.0
pandas==2.2.2
langchain==0.2.1
langchain-huggingface==0.0.4
huggingface-hub==0.22.2
ydata-profiling==4.6.4
matplotlib==3.8.4
seaborn==0.13.2
scikit-learn==1.4.2
python-dotenv==1.0.1
numpy==1.26.4
